quic: make alpn default to quic.alpnVersion
Currently, in doc/api/quic.md the first quic server example does not
show the alpn option being passed to the listen function. There is an
explicit check in core.js where this option is checked and invalid type
error will be thrown if it not specified, or not of type string..

This commit suggests defaulting to the alpnVersion the alpn option is
not specified.

PR-URL: https://github.com/nodejs/quic/pull/105
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Anna Henningsen <anna@addaleax.net>