quic: make GetAddress take host and host_len
Currently, GetAddress takes a char** and assigns to a stack allocated
char buffer. This will cause logging in QuicSocket::SendPacket to look
like this:
QuicSocket (0:3) Sending to  at port 1234

This should instead be something like:
QuicSocket (0:3) Sending to 127.0.0.1 at port 1234

This commit changes the host parameter of GetAddress to be a char* and,
adds a host_len parameter which should be the length of the char array.
This seems to be more efficient than allocating/copying but I'd be happy
to revisit if others feel differently.

PR-URL: https://github.com/nodejs/quic/pull/127
Reviewed-By: James M Snell <jasnell@gmail.com>