quic: make `maxConnectionsPerHost` works
It seems that `Hash` class for `unorderred_map` is used to decide which
bucket to put the key. But keys that are different pointers with the
same hash will still be regarded as different keys so that we still need
to specify a `std::equal_to` to compare the key correctly.

However, when `maxConnectionsPerHost` is exceeded, the new session will
still be emitted on the server side and won't be destroied automatically.
We need to figure out the reason and fix it.

PR-URL: https://github.com/nodejs/quic/pull/186
Reviewed-By: Anna Henningsen <anna@addaleax.net>