doc,quic: update BigInt type to lowercase
Currently the doc target fails with the following error:
Error: Unrecognized type: 'BigInt'.
Please, edit the type or update the 'tools/doc/type-parser.js'.
    at /quic/quic/tools/doc/type-parser.js:199:15
    at Array.forEach (<anonymous>)
    at Object.toLink (/quic/tools/doc/type-parser.js:175:13)
    at /quic/quic/tools/doc/html.js:164:59
    at Array.forEach (<anonymous>)
    at linkJsTypeDocs (/quic/quic/tools/doc/html.js:163:19)
    at /quic/tools/doc/html.js:111:23

This commit updates the types of BigInt to be lowercase which is how it
is defined in type-parser.js. I would have thought using 'BigInt' would
have been better but there might be a reason for this that I'm unware of
('bigint' is used in other files).

With this commit the build passes and I'll look into why this type is
lowercase and suggest fixing it in upstream/master.

PR-URL: https://github.com/nodejs/quic/pull/385
Reviewed-By: James M Snell <jasnell@gmail.com>